change return type of getRemainTime from int to long

+++ src/main/java/it/dpg/minigames/molegame/model/Timer.java
@@ -15,8 +15,10 @@ public interface Timer {
 
     /**
      * get the remain time
+     *
+     * @return remain time
      */
-    int getRemainTime();
+    long getRemainTime();
